adding ApplicationBeanConfig not using reflections at all for scanning but only jaxrs Application as alternative to BeanConfig
